+++ packages/types/src/modules/auth-types.ts
@@ -11,3 +11,22 @@ export interface LoginWithMagicLinkConfiguration {
    */
   showUI?: boolean;
 }
+
+export interface RegisterWithWebAuthnConfiguration {
+  /**
+   * The username of the user attempting to register.
+   */
+  username: string;
+
+  /**
+   * The nickname that the user attempts to set to this webauthn device.
+   */
+  nickname?: string;
+}
+
+export interface LoginWithWebAuthnConfiguration {
+  /**
+   * The username of the user attempting to register.
+   */
+  username: string;
+}